% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asum.R
\name{asum}
\alias{asum}
\title{Aspect Sentiment Unification Model}
\usage{
asum(sfm, sentimentLexicon = NULL, numIters = 3, numSentiLabs = 2,
  numTopics = 5, alpha = 0.1, gamma = 1, betaVec = c(0.001, 0.9, 0.1),
  excludeNeutral = TRUE)
}
\arguments{
\item{sfm}{A \pkg{\link{quanteda}} dfm object split into sentences. To split a corpus, see
\link{corpus_sentences}.}

\item{sentimentLexicon}{A \pkg{\link{quanteda}} dictionary object containing seed words for
the sentiments to be estimated. Prior sentiment categories for sentences are set if to a category
if one or more words in the sentence are found in that category and none were found in other 
categories. Therefore a larger dictionary does not automatically provide a better initialisation.
The best dictionary i one without words that could possibly be ambiguous, especially if an 
asymmetrical beta is used. See \code{\link{paradigm}} for the dictionary proposed by the 
authors of ASUM.}

\item{numIters}{Integer. The number of iterations to run.}

\item{numSentiLabs}{Integer. The number of sentiment labels to estimate. This is overridden if
a dictionary with seed words is included. If you wish to add a neutral category, set 
\code{excludeNeutral = FALSE}. By default, ASUM includes no neutral category.}

\item{numTopics}{Integer. The number of topics to be estimated.}

\item{alpha}{Numeric. The dirichlet parameter for the sentiment distribution prior. Defaults to 0.1.}

\item{gamma}{Numeric. The dirichlet parameter for the topic distribution prior. Defaults to 1.}

\item{betaVec}{Numeric(3). This vector asks for three elements, namely:
\itemize{
  \item{\code{betaVec[1]}}{The beta parameter to be taken for all words}
  \item{\code{betaVec[2]}}{The multiplier for the same sentiment category as the word has in
  the included sentiment lexicon. If no sentiment lexicon is included or a word is not in the 
  lexicon, this argument is ignored.}
  \item{\code{betaVec[3]}}{The multiplier for the other sentiment category/categories as the word
  has in the included sentiment lexicon is included or a word is not in the lexicon, this 
  argument is ignored.}
}
Beta can be set symmetrically as \code{betaVec = c(beta,1,1)} or asymmetrically by setting for
example \code{betaVec = c(beta,1,0.01)}. If \code{betaVec[3] = 0}, the calculations are streamlined
by setting posterior probabilities to 0 for sentiment categories other than the one a sentence is
initialised in if it was initialised from the sentiment lexicon. Potentially, his slightly deviates 
from the posterior distribution, especially if combined with larger sentiment dictionaries. It is
more efficient, but not advised for every application. Choose the amount of asymmetry carefully.}

\item{excludeNeutral}{Boolean. This variable instructs whether an extra category for neutral
sentences should be added to the categories provided by the sentiment dictionary.}
}
\value{
An ASUM.result object containing the parameters from the estimated model.
}
\description{
This method performs an estimation of the Aspect Sentiment Unification Model, which performs
a task similar to \code{\link{jst}}, but categorises sentences rather than words. The estimation is
implemented with a Gibbs sampler.
}
