% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_sentences.R
\name{corpus_sentences}
\alias{corpus_sentences}
\title{Split a corpus into sentences}
\usage{
corpus_sentences(corpus)
}
\arguments{
\item{corpus}{A quanteda corpus object}
}
\value{
A quanteda corpus object with sentences split. This has the following changes:
        \itemize {
         \item{sentenceNo}{The number of the sentences within documents. The are kept in
         order of appearance.}
         \item{docId}{The old id variable (document id) is replicated for every sentence and
         stored under the name docId. This way the documents can be combined again later if
         necessary (as it is for sentence LDA or ASUM)}
         \item{id}{This is the new unique 'document' id within the corpus structure. It is
         structured as follows: [docID]-[sentenceNo]. This will from here onward be treated
         as the document ID, since the tokens and dfm methods do not discriminated between 
         documents and sentences as elements of a corpus object.}
        }
}
\description{
This method splits a \pkg{quanteda} corpus object into sentences. Each sentence is
then returned as a document. This method is an addition to the possibility to split
sentences using the tokens method. The resulting tokens object can then no longer be
split to words, whereas the resulting corpus from this method can be turned into a 
tokens object (and consequently a dfm) in which the sentences remain split.
}
