% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst.R
\name{is.JST.result}
\alias{is.JST.result}
\title{Check if an object is a JST.result object}
\usage{
is.JST.result(x)
}
\arguments{
\item{x}{object}
}
\value{
Boolean. True if x is a JST.result object.
}
\description{
Check if an object is a JST.result object
}
