% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst_reversed.R
\name{is.JST_reversed.result}
\alias{is.JST_reversed.result}
\title{Check if an object is a JST_reversed.result object}
\usage{
is.JST_reversed.result(x)
}
\arguments{
\item{x}{object}
}
\value{
Boolean. True if x is a JST_reversed.result object.
}
\description{
Check if an object is a JST_reversed.result object
}
