% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{is.LDA.result}
\alias{is.LDA.result}
\title{Check if an object is a LDA.result object}
\usage{
is.LDA.result(x)
}
\arguments{
\item{x}{object}
}
\value{
Boolean. True if x is a LDA.result object.
}
\description{
Check if an object is a LDA.result object
}
