% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentenceLDA.R
\name{is.sLDA.result}
\alias{is.sLDA.result}
\title{Check if an object is a sLDA.result object}
\usage{
is.sLDA.result(x)
}
\arguments{
\item{x}{object}
}
\value{
Boolean. True if x is a LDA.result object.
}
\description{
Check if an object is a sLDA.result object
}
