% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst.R
\name{jst}
\alias{jst}
\title{Run a Joint Sentiment Topic model}
\usage{
jst(dfm, sentiLexInput = list(), numSentiLabs = 3, numTopics = 10,
  numIters = 3, updateParaStep = -1, alpha = -1, beta = -1,
  gamma = -1)
}
\arguments{
\item{dfm}{A quanteda dfm object}

\item{sentiLexInput}{Optional: A quanteda dictionary object for semi-supervised learning}

\item{numSentiLabs}{Integer, the number of sentiment labels (defaults to 3)}

\item{numTopics}{Integer, the number of topics (defaults to 10)}

\item{numIters}{Integer, the number of iterations (defaults to ???)}

\item{updateParaStep}{Integer. The number of iterations between optimizations of hyperparameter alpha}

\item{alpha}{Double, hyperparameter for (defaults to)}

\item{beta}{Double, hyperparameter for (defaults to)}

\item{gamma}{Double, hyperparameter for (defaults to)}
}
\value{
A JST.result object containing a data.frame for each estimated parameter
}
\description{
Run a Joint Sentiment Topic model
}
