% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst.R
\docType{class}
\name{JST.result-class}
\alias{JST.result-class}
\alias{JST.result}
\title{JST results object}
\description{
Contains estimated Joint Sentiment Topic model
}
\section{Slots}{

\describe{
\item{\code{pi}}{Document-level sentiment estimates}

\item{\code{theta}}{Document-level sentitopic estimates}

\item{\code{phi}}{Word-level sentitopic estimates}

\item{\code{phi.termScores}}{Word-level term scores (suboptimal calculation, only useful for smaller models)}

\item{\code{numTopics}}{Number of topics}

\item{\code{numSentiments}}{Number of sentiment categories}

\item{\code{docvars}}{Document-level metadata from the quanteda object used as input}
}}

