% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst_reversed.R
\name{jst_reversed}
\alias{jst_reversed}
\title{Run a reversed Joint Sentiment Topic model}
\usage{
jst_reversed(dfm, sentiLexInput = list(), numSentiLabs = 3,
  numTopics = 10, numIters = 3, updateParaStep = -1, alpha = -1,
  beta = -1, gamma = -1)
}
\arguments{
\item{dfm}{A quanteda dfm object}

\item{sentiLexInput}{Optional: A quanteda dictionary object for semi-supervised 
learning}

\item{numSentiLabs}{Integer, the number of sentiment labels (defaults to 3)}

\item{numTopics}{Integer, the number of topics (defaults to 10)}

\item{numIters}{Integer, the number of iterations (defaults to ???)}

\item{updateParaStep}{Integer. The number of iterations between optimizations 
of hyperparameter alpha}

\item{alpha}{Double, hyperparameter for (defaults to)}

\item{beta}{Double, hyperparameter for (defaults to)}

\item{gamma}{Double, hyperparameter for (defaults to)}
}
\value{
A JST_reversed.result object containing a data.frame for each estimated 
parameter
}
\description{
Note: For explanations on the reversed JST, send me an email, although it runs 
it is still work in progress.
}
