% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst_reversed.R
\docType{class}
\name{JST_reversed.result-class}
\alias{JST_reversed.result-class}
\alias{JST_reversed.result}
\title{Reverse JST results object}
\description{
Contains estimated reversed Joint Sentiment Topic model, see Details for model description.
}
\section{Slots}{

\describe{
\item{\code{pi}}{Document-level topic estimates}

\item{\code{theta}}{Document-level sentitopic estimates}

\item{\code{phi}}{Word-level sentitopic estimates}

\item{\code{phi.termScores}}{Word-level term scores (suboptimal calculation, only useful for smaller models)}

\item{\code{numTopics}}{Number of topics}

\item{\code{numSentiments}}{Number of sentiment categories}

\item{\code{docvars}}{Document-level metadata from the quanteda object used as input}
}}

