% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{lda}
\alias{lda}
\title{Latent Dirichlet Allocation}
\usage{
lda(dfm, numTopics, numIters, alpha, beta)
}
\arguments{
\item{dfm}{A quanteda dfm object}

\item{numTopics}{Integer, the number of topics}

\item{numIters}{Integer, the number of iterations}

\item{alpha}{Double, hyperparameter for the dirichlet prior of the topic
distribution}

\item{beta}{Double, hyperparameter for the dirichlet prior of the words
distribution}
}
\value{
An LDA.result object containing a data.frame for each estimated parameter
}
\description{
Gibbs sampling estimation of Latent Dirichlet Allocation (LDA). Under development,
so no defaults have been set.
}
