% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{paradigm}
\alias{paradigm}
\title{Paradigm word list (Lin and He, 2009, p.379)}
\usage{
paradigm()
}
\value{
A quanteda dictionary2 object containing the paradigm word list
}
\description{
Returns the paradigm word list proposed by Lin and He. These words create groupings
of positive and negative words which are considered to be domain-independent. However,
as always, caution is warranted. The words are mostly suited for review analysis. Review
the words before applying the dictionary.
}
\note{
Lin, C., & He, Y. (2009). Joint sentiment/topic model for sentiment analysis. 
In Proceedings of the 18th ACM conference on Information and knowledge management 
(pp. 375-384). ACM.
}
