% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst.R
\name{plot.JST.result}
\alias{plot.JST.result}
\title{Plot document sentiment scores}
\usage{
\method{plot}{JST.result}(x, sentiment1, sentiment2, colourBy = NULL)
}
\arguments{
\item{x}{A JST.result object}

\item{sentiment1}{Integer}

\item{sentiment2}{Integer}

\item{colourBy}{(Optional) Character. The name of a variable from the docvars of the original tokens file}
}
\value{
A ggplot2 scatterplot of the document sentiments
}
\description{
Plot document sentiment scores
}
