\name{rJST-package}
\alias{rJST-package}
\alias{rJST}
\docType{package}
\title{
\packageTitle{rJST}
}
\description{
\packageDescription{rJST}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rJST}
\packageIndices{rJST}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rJST}

Maintainer: \packageMaintainer{rJST}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
