% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentenceLDA.R
\name{sentencelda}
\alias{sentencelda}
\title{Sentence Latent Dirichlet Allocation}
\usage{
sentencelda(dfm, numTopics, numIters, alpha, beta)
}
\arguments{
\item{dfm}{A quanteda dfm object, split into sentences}

\item{numTopics}{Integer, the number of topics}

\item{numIters}{Integer, the number of iterations}

\item{alpha}{Double, hyperparameter for the dirichlet prior of the topic
distribution}

\item{beta}{Double, hyperparameter for the dirichlet prior of the words
distribution}
}
\value{
An sLDA.result object containing a data.frame for each estimated parameter
}
\description{
Gibbs sampling estimation of sentence Latent Dirichlet Allocation (sLDA). Note that the
primary argument to this method is a dfm object in which the documents are expected to
be split into sentences. This can be done earlier in the process, using the method 
\code{\link{corpus_sentences}}. There is no check for this requirement, so keep in mind
to do it yourself.
}
\seealso{
\code{\link{corpus_sentences}}
}
