% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asum.R
\name{tidy.ASUM.result}
\alias{tidy.ASUM.result}
\title{Tidy ASUM results}
\usage{
\method{tidy}{ASUM.result}(x, parameter = NULL)
}
\arguments{
\item{x}{A ASUM.result object}

\item{parameter}{Character. The parameter to be tidied and returned. 
Note that no default is set.}
}
\value{
A tidy data.frame.
}
\description{
This method tidies up the results object for the selected parameter
and returns a data.frame that conforms to the standards from the
R tidyverse. See \pkg{broom} for the generic tidy method.
}
