% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst.R
\name{top20words}
\alias{top20words}
\title{Show the top 20 words for a topic/sentiment combination}
\usage{
top20words(x, topic, sentiment)
}
\arguments{
\item{x}{A JST.result object}

\item{topic}{Integer}

\item{sentiment}{Integer}
}
\value{
A CharacterVector containing the 20 top words of the topic/sentiment combination
}
\description{
Show the top 20 words for a topic/sentiment combination
}
