% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{wordstemdict}
\alias{wordstemdict}
\title{Wordstem a quanteda Dictionary2 object}
\usage{
wordstemdict(dict)
}
\arguments{
\item{dict}{A quanteda dictionary2 object}
}
\value{
A quanteda dictionary2 object with all elements stemmed
}
\description{
Applies Porter stemming to the words in a quanteda dictionary and then removes 
duplicates, if the stemming created these.
}
