% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rJST.R
\docType{package}
\name{rJST-package}
\alias{rJST}
\alias{rJST-package}
\title{rJST: Joint Sentiment Topic modelling for R}
\description{
rJST performs Joint Sentiment Topic modelling and includes a
reversed model (for info on the latter, email me) and several
methods to evaluate the results and make them more accessible.
All models take \pkg{quanteda} dfm objects as inputs. Furthermore,
all models have a defined tidy method. See \pkg{broom} for more
information.
}
\author{
\strong{Maintainer}: Max Boiten \email{max_boiten@live.nl}

Other contributors:
\itemize{
  \item Chenghua Lin [copyright holder]
  \item Yulan He [copyright holder]
}

}
