% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst.R, R/jst_reversed.R
\name{top20words}
\alias{top20words}
\alias{top20words}
\title{Show the top 20 words for a topic/sentiment combination}
\usage{
top20words(x, topic, sentiment, termScores = TRUE)

top20words(x, topic, sentiment, termScores = TRUE)
}
\arguments{
\item{x}{A JST.result object}

\item{topic}{Integer}

\item{sentiment}{Integer}

\item{termScores}{Boolean. TRUE is you wish to use term scores (Lafferty and Blei, 2009)
rather than the phi parameter as estimated by JST. Defaults to TRUE.}

\item{x}{A JST.result object}

\item{topic}{Integer}

\item{sentiment}{Integer}

\item{termScores}{Boolean. TRUE is you wish to use term scores (Lafferty and Blei, 2009)
rather than the phi parameter as estimated by JST. Defaults to TRUE.}
}
\value{
A CharacterVector containing the 20 top words of the topic/sentiment combination

A CharacterVector containing the 20 top words of the topic/sentiment combination
}
\description{
Show the top 20 words for a topic/sentiment combination

Show the top 20 words for a topic/sentiment combination
}
