% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst.R
\name{topNwords}
\alias{topNwords}
\title{Show the top N words for a topic/sentiment combination}
\usage{
topNwords(x, topic, sentiment, N)
}
\arguments{
\item{x}{A JST.result object}

\item{topic}{Integer}

\item{sentiment}{Integer}

\item{N}{Integer, the number of words to be returned}
}
\value{
A CharacterVector containing the N top words of the topic/sentiment combination
}
\description{
Show the top N words for a topic/sentiment combination
}
