% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst.R
\name{topNwords}
\alias{topNwords}
\title{Show the top N words for a topic/sentiment combination}
\usage{
topNwords(x, topic, sentiment, N, termScores = TRUE)
}
\arguments{
\item{x}{A JST.result object}

\item{topic}{Integer}

\item{sentiment}{Integer}

\item{N}{Integer, the number of words to be returned}

\item{termScores}{Boolean. TRUE is you wish to use term scores (Lafferty and Blei, 2009)
rather than the phi parameter as estimated by JST. Defaults to TRUE.}
}
\value{
A CharacterVector containing the N top words of the topic/sentiment combination
}
\description{
Show the top N words for a topic/sentiment combination
}
